Update dependencies (guava, etc) (google/MOE#41)
Remove other snapshot dependencies in `auto-value` and `auto-factory`

When attempting to execute the built `moe` binary after running `make-binary.sh` it exits with an exception.
It exits with a `Caused by: java.lang.ClassNotFoundException: com.google.auto.factory.internal.Preconditions`.
This removes the snapshot dependencies from the projecgt and the snapshots repository.